made filterEmails a utils function
