+++ client/pages/components/EmailList.js
@@ -10,6 +10,7 @@ import React, { useState, useEffect } from "react";
 import axios from "axios";
 import EmailHeader from "./EmailHeader";
 import defaults from "../../utils/defaults";
+import filterEmails from "../../utils/filterEmails";
 
 /**
  * This component retrieves and renders a list of emails.
@@ -21,7 +22,7 @@ export default function EmailListOld({ isSentPage }) {
   const [searchVal, setSearchVal] = useState("");
 
   useEffect(() => {
-    setShowEmailIndices(getIndices(emails, searchVal));
+    setShowEmailIndices(filterEmails(emails, searchVal));
   }, [searchVal]);
 
   // get all the emails
@@ -91,48 +92,4 @@ export default function EmailListOld({ isSentPage }) {
       </div>
     </div>
   );
-}
-
-/**
- * Find the indices of emails matching a search term
- * @param emails    array of emails to search through and reference the index of
- * @param searchVal the value we are searching for in the emails array
- * @returns         an array of indices that match the search term
- */
-function getIndices(emails, searchVal) {
-  // if no search term, then render all
-  if (searchVal.length === 0) {
-    return [...Array(emails.length).keys()];
-  }
-
-  const indices = [];
-
-  emails.forEach((email, index) => {
-    // subject search
-    if (email.subject.toLowerCase().includes(searchVal.toLowerCase())) {
-      indices.push(index);
-      return;
-    }
-
-    // body search
-    if (email.body.toLowerCase().includes(searchVal.toLowerCase())) {
-      indices.push(index);
-      return;
-    }
-
-    // address search
-    const addresses = [
-      email.from.email,
-      ...email.to.map((to) => to.email),
-      ...email.cc.map((cc) => cc.email),
-    ];
-    for (const addr of addresses) {
-      if (addr.toLowerCase().includes(searchVal.toLowerCase())) {
-        indices.push(index);
-        return;
-      }
-    }
-  });
-
-  return indices;
 }
